Version 1 of Pvalue Testing for all indicators
It reads in indicators which are proportions with targets and identifies the pvalues that the indicators came from a distribution with that target. Positive or negative deviance change the context of the tests flipping it from lower or upper tail tests. This is a framing that can make the results more politically viabile.

Important related logics are contained in the database SQL objects. Results are saved back into the SQL database.

Remove the rmarkdown report as it's no longer required.
The library file might need to change.